Resolve property naming bug preventing attribute from being added to firestore.

+++ functions/src/index.ts
@@ -12,46 +12,49 @@ const db = admin.firestore();
 ///   3. Merge the value which we are adding to the appropriate nested field.
 exports.addToNewFlyFormTemplate = functions.firestore.document('/new_fly_form_incoming/{docId}')
   .onWrite(async (change, context) => {
-    const newFlyFormTemplate = 'new_fly_form';
+    if (change.after.data()) {
+      const newFlyFormTemplate = 'new_fly_form';
 
-    const prevDoc = await db.collection(newFlyFormTemplate)
-      .orderBy('last_modified', 'desc').limit(1).get();
-    const prevTemplateData = prevDoc.docs[0].data();
-    prevTemplateData.last_modified = new Date();
+      const prevDoc = await db.collection(newFlyFormTemplate)
+        .orderBy('last_modified', 'desc').limit(1).get();
+      const prevTemplateData = prevDoc.docs[0].data();
+      prevTemplateData.last_modified = new Date();
 
-    // Enforced by our security rules, docId is this user's id.
-    prevTemplateData.updated_by = change.after.id;
+      // Enforced by our security rules, docId is this user's id.
+      prevTemplateData.updated_by = change.after.id;
 
-    const newFormTemplateDoc = await db.collection(newFlyFormTemplate).add(prevTemplateData);
+      const newFormTemplateDoc = await db.collection(newFlyFormTemplate).add(prevTemplateData);
 
-    if (change.after.data()?.attributes) { // Path to add attributes.
-      const updateData = change.after.data()?.attributes;
-      await db.collection(newFlyFormTemplate).doc(newFormTemplateDoc.id)
-        .update({
-          ['attributes.' + Object.keys(updateData)[0]]:
-            admin.firestore.FieldValue.arrayUnion(Object.values(updateData)[0]),
-        });
+      if (change.after.data()?.attributes) { // Path to add attributes.
+        const updateData = change.after.data()?.attributes;
+        await db.collection(newFlyFormTemplate).doc(newFormTemplateDoc.id)
+          .update({
+            ['attributes.' + Object.keys(updateData)[0]]:
+              admin.firestore.FieldValue.arrayUnion(Object.values(updateData)[0]),
+          });
 
-      return db.collection('new_fly_form_incoming').doc(change.after.id).delete();
-    }
-    else if (change.after.data()?.materials) { // Path to add materials.
-      // change.after.data()?.materials looks something like:
-      // {beads: {color: green}}
-      const updateData : {[key: string]: {[val: string]: string}} = change.after.data()?.materials;
+        return db.collection('new_fly_form_incoming').doc(change.after.id).delete();
+      }
+      else if (change.after.data()?.materials) { // Path to add materials.
+        // change.after.data()?.materials looks something like:
+        // {beads: {color: green}}
+        const updateData: { [key: string]: { [val: string]: string } } = change.after.data()?.materials;
 
-      // keyTL would be 'beads'
-      const keyTL : string =  Object.keys(updateData)[0];
-      //  keyBL would be 'color'
-      const keyBL : string = String(Object.keys(updateData[keyTL])[0]);
-      const val : string = updateData[keyTL][keyBL];
+        // keyTL would be 'beads'
+        const keyTL: string = Object.keys(updateData)[0];
+        //  keyBL would be 'color'
+        const keyBL: string = String(Object.keys(updateData[keyTL])[0]);
+        const val: string = updateData[keyTL][keyBL];
 
-      await db.collection(newFlyFormTemplate).doc(newFormTemplateDoc.id)
-        .update({
-          ['materials.' + keyTL + '.' + keyBL]:
-            admin.firestore.FieldValue.arrayUnion(val),
-        });
+        await db.collection(newFlyFormTemplate).doc(newFormTemplateDoc.id)
+          .update({
+            ['materials.' + keyTL + '.' + keyBL]:
+              admin.firestore.FieldValue.arrayUnion(val),
+          });
 
-      return db.collection('new_fly_form_incoming').doc(change.after.id).delete();
+        return db.collection('new_fly_form_incoming').doc(change.after.id).delete();
+      }
+      else return; // Not a used path
     }
     else {
       // Path to prevent infinite loop.
